refactor: move contexts.rs test block to sibling tests.rs (#1578)
Splits frontend/src/contexts.rs into a directory module (mod.rs +
tests.rs) following the db/src/books/tests.rs pattern from rule
03-unit-testing.md. Pure file move: the inline 13-test, 186-line
mod tests block becomes contexts/tests.rs unchanged; mod.rs keeps
`#[cfg(test)] mod tests;` in its place. No logic or assertion changes.
